fix: various anoncreds revocation fixes
