+++ packages/anoncreds/src/index.ts
@@ -13,3 +13,4 @@ export { AnonCredsApi } from './AnonCredsApi'
 export * from './AnonCredsApiOptions'
 export { generateLegacyProverDidLikeString } from './utils/proverDid'
 export * from './utils/legacyIndyIdentifiers'
+export { assertBestPracticeRevocationInterval } from './utils/revocationInterval'
