+++ packages/anoncreds/src/utils/index.ts
@@ -3,7 +3,8 @@ export { sortRequestedCredentialsMatches } from './sortRequestedCredentialsMatch
 export { assertNoDuplicateGroupsNamesInProofRequest } from './hasDuplicateGroupNames'
 export { areAnonCredsProofRequestsEqual } from './areRequestsEqual'
 export { downloadTailsFile } from './tails'
-export { assertRevocationInterval } from './revocationInterval'
+export { assertBestPracticeRevocationInterval } from './revocationInterval'
+export { getRevocationRegistriesForRequest, getRevocationRegistriesForProof } from './getRevocationRegistries'
 export { encodeCredentialValue, checkValidCredentialValueEncoding } from './credential'
 export { IsMap } from './isMap'
 export { composeCredentialAutoAccept, composeProofAutoAccept } from './composeAutoAccept'
